# Fix benchmark to reuse parsers