--- a/benches/benchmark.rs
+++ b/benches/benchmark.rs
@@ -1,5 +1,7 @@
 #![feature(test)]
 
+#[macro_use]
+extern crate lazy_static;
 extern crate test;
 extern crate uap_rust;
 extern crate woothee;
@@ -8,18 +10,21 @@ use test::Bencher;
 use uap_rust::parser as uap;
 use woothee::parser as woo;
 
+lazy_static! {
+    static ref UAP_PARSER: uap::Parser = uap::Parser::new().unwrap();
+    static ref WOO_PARSER: woo::Parser = woo::Parser::new();
+}
+
 fn b_uap() {
-    let parser = uap::Parser::new().unwrap();
-    let _ = parser.parse("Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 6.1; Trident/4.0)".to_string());
-    let _ = parser.parse("Twitterbot/1.0".to_string());
-    let _ = parser.parse("Mozilla/5.0 (compatible; MSIE 9.0; Windows NT 6.1; Trident/5.0; Xbox)".to_string());
+    let _ = UAP_PARSER.parse("Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 6.1; Trident/4.0)".to_string());
+    let _ = UAP_PARSER.parse("Twitterbot/1.0".to_string());
+    let _ = UAP_PARSER.parse("Mozilla/5.0 (compatible; MSIE 9.0; Windows NT 6.1; Trident/5.0; Xbox)".to_string());
 }
 
 fn b_woothee() {
-    let parser = woo::Parser::new();
-    let _ = parser.parse("Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 6.1; Trident/4.0)");
-    let _ = parser.parse("Twitterbot/1.0");
-    let _ = parser.parse("Mozilla/5.0 (compatible; MSIE 9.0; Windows NT 6.1; Trident/5.0; Xbox)");
+    let _ = WOO_PARSER.parse("Mozilla/4.0 (compatible; MSIE 8.0; Windows NT 6.1; Trident/4.0)");
+    let _ = WOO_PARSER.parse("Twitterbot/1.0");
+    let _ = WOO_PARSER.parse("Mozilla/5.0 (compatible; MSIE 9.0; Windows NT 6.1; Trident/5.0; Xbox)");
 }
 
 #[bench]