Initialisation of parser might be quite expensive, and, obviously, in real use-case it's meant to be reused, so this commit moves parser inits into lazy_static! block for more fair comparison.

Before:
```
test bench_stabilizer ... bench:          30 ns/iter (+/- 10)
test smoke_uap        ... bench: 229,254,862 ns/iter (+/- 14,099,776)
test smoke_woothee    ... bench:       8,678 ns/iter (+/- 1,500)
```

After:
```
test bench_stabilizer ... bench:          24 ns/iter (+/- 2)
test smoke_uap        ... bench:     352,355 ns/iter (+/- 68,453)
test smoke_woothee    ... bench:       8,657 ns/iter (+/- 1,178)
```

Woothee is still much faster and hasn't changed numbers too much, but UAP numbers became ~650x better as they now measure just UA parsing and not loading the database.