@hhatto 2018-06-26
@RReverser 

Thank you for contribution.
LGTM 👍